fix: Seperate thunks
