+++ redux/slice.ts
@@ -1,6 +1,4 @@
-import { authServiceLogin } from '@/services/auth-service';
 import { createSlice } from '@reduxjs/toolkit';
-import { Dispatch } from 'redux';
 
 export type User = {
   accessToken: string;
@@ -41,13 +39,3 @@ export const { reducer, actions } = createSlice({
 
 export default reducer;
 export const { setAccessToken, setUserEmail, setUserPhotoURL } = actions;
-
-export function requestLogin(providerName: 'GitHub' | 'Google') {
-  return async (dispatch: Dispatch) => {
-    const data = await authServiceLogin(providerName);
-
-    dispatch(setAccessToken(data.credential?.accessToken));
-    dispatch(setUserEmail(data.user?.email));
-    dispatch(setUserPhotoURL(data.user?.photoURL));
-  };
-}
